+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## ALPHA Build #2 (v0.1.0)
+August 19, 2022
+
+* Changed added strings to f strings
+* New 'Internet' commands (Refer to the help command)
+* Fixed Issue [#1](https://github.com/KnowingVR/Flock/issues/1)
+___
 ### Repository Update
 
 August 18, 2022